+++ lib/go/model_field_mapping_v2.go
@@ -18,6 +18,8 @@ import (
 type FieldMappingV2 struct {
 	// A user specified string that is a name for this mapping
 	Name string `json:"name,omitempty"`
+	// A boolean that determines whether to drop all fields in this document. If set, input and output fields should not be set
+	IsDropAllFields bool `json:"is_drop_all_fields,omitempty"`
 	// A List of InputField for this mapping
 	InputFields []InputField `json:"input_fields,omitempty"`
 	// An OutputField for this mapping